Topテンプレートにpropsを渡せるようにした
