+++ components/organisms/Section/index.stories.js
@@ -2,39 +2,8 @@ import HeaderNav from '@/components/organisms/Section/HeaderNav.vue'
 import PortfolioCardList from '@/components/organisms/Section/PortfolioCardList.vue'
 import SiteFooter from '@/components/organisms/Section/SiteFooter.vue'
 
-function generateSampleData() {
-  const samples = []
-  const url = 'https://www.y-mix.dev/'
-  const imageUrl = ''
-  const tagSample = [
-    { text: 'Go' },
-    { text: 'Apache Spark' },
-    { text: 'Twitter API' },
-    { text: 'ShellScript' },
-    { text: 'Flask' },
-    { text: 'jQuery' },
-    { text: 'Bulma' },
-    { text: 'nginx' },
-    { text: 'Jenkins' },
-    { text: 'BigQuery' },
-    { text: 'Ionic' },
-    { text: 'Bootstrap' },
-    { text: 'iOS(Objective-C)' },
-    { text: 'Android(Java)' },
-    { text: 'Unity(JS, C#)' }
-  ]
-  for (let i = 0; i < 20; i++) {
-    samples.push({
-      url,
-      imageUrl,
-      title: 'Sample ' + String(i),
-      skills: tagSample
-    })
-  }
-
-  return samples
-}
-
+// サンプルデータの読み込み
+import { generateSampleData } from '@/.storybook/sampleData.js'
 const sampleData = generateSampleData()
 
 export default {
