+++ components/templates/index.stories.js
@@ -1,10 +1,15 @@
 import Top from '@/components/templates/Top.vue'
 
+// サンプルデータの読み込み
+import { generateSampleData } from '@/.storybook/sampleData.js'
+const sampleData = generateSampleData()
+
 export default {
   title: '/templates/TopPage'
 }
 
 export const top = () => ({
   components: { Top },
-  template: `<top />`
+  data: () => ({ sampleData }),
+  template: `<top :portfolio-list-prop="sampleData"/>`
 })
